requirements.txt
fastapi==0.128.0
uvicorn==0.39.0
python-multipart==0.0.9

pydantic==2.12.5
numpy==1.26.4
pandas==2.3.3
scikit-learn==1.6.1

Pillow==10.4.0
matplotlib==3.9.4

torch==2.3.1
torchvision==0.18.1
torchaudio==2.3.1

mlflow==3.1.4
boto3==1.35.99
psycopg2-binary==2.9.10

prometheus-client==0.22.0
prometheus-fastapi-instrumentator==7.1.0

requests==2.32.5
PyYAML==6.0.3

pytest==8.4.2
ruff==0.14.10
tabulate==0.9.0